# infrastructure/swagger/openapi.yml
openapi: 3.0.3
info:
  title: Smart Banking System API
  description: |
    Complete REST API for the TT-BANK Smart Banking System.

    **Architecture:** Microservices — all requests enter through the API Gateway on port 8080.
    The gateway validates the JWT Bearer token before routing to the appropriate backend service.

    **Authentication flow:**
    1. `POST /api/v1/auth/register` — create account
    2. `POST /api/v1/auth/verify-email` — verify OTP sent by email
    3. `POST /api/v1/auth/login` — receive `accessToken` and `refreshToken`
    4. Include `Authorization: Bearer <accessToken>` on all subsequent requests

    **Currency:** Primary currency is **XAF** (Central African CFA franc). Also supports USD, EUR, GBP.

    **Base URL (local dev):** `http://localhost:8080`
    **Base URL (production):** `https://yourdomain.com`
  version: "1.0.0"
  contact:
    name: Smart Banking Dev Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: https://yourdomain.com
    description: Production (replace with actual domain)

# =============================================================
# Security scheme — JWT Bearer token (HS256)
# =============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by POST /api/v1/auth/login.
        Include as: Authorization: Bearer <token>

  # =============================================================
  # Reusable schemas — built from actual Java DTOs
  # =============================================================
  schemas:
...
    AuditEventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        eventDomain:
          $ref: '#/components/schemas/EventDomain'
        eventType:
          $ref: '#/components/schemas/EventType'
        referenceCode:
          type: string
          example: "TXN-20240601-ABC123"
        actorEmail:
          type: string
          example: "jeanpierre@example.com"
        summary:
          type: string
          example: "Transfer of 5000.00 XAF from WLT-001 to WLT-002 completed"
        rawPayload:
          type: string
          description: Full JSON event payload as stored
          example: '{"amount":5000.00,"currency":"XAF",...}'
        eventTimestamp:
          type: string
          format: date-time
          example: "2024-06-01T14:30:01"
        receivedAt:
          type: string
          format: date-time
          example: "2024-06-01T14:30:01.123"
    EventDomain:
      type: string
      enum: [WALLET, TRANSACTION, MERCHANT, SAVINGS]
    EventType:
      type: string
      enum:
        - WALLET_CREATED
        - WALLET_FUNDED
        - WALLET_WITHDRAWN
        - TRANSACTION_INITIATED
        - TRANSACTION_COMPLETED
        - TRANSACTION_FAILED
        - MERCHANT_REGISTERED
        - MERCHANT_PAYMENT_INITIATED
        - MERCHANT_PAYMENT_COMPLETED
        - MERCHANT_PAYMENT_FAILED
        - SAVINGS_GROUP_CREATED
        - SAVINGS_MEMBER_JOINED
        - SAVINGS_CONTRIBUTION_MADE
        - SAVINGS_PAYOUT_PROCESSED
        - UNKNOWN

    # ── Shared ────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed: email must be a valid email address"
        timestamp:
          type: string
          format: date-time
          example: "2024-06-01T14:30:00"
        path:
          type: string
          example: "/api/v1/auth/register"
  responses:
    Unauthorized:
      description: Missing or invalid JWT Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Valid token but insufficient role (e.g. non-ADMIN accessing /api/v1/admin/**)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# =============================================================
# API Paths
# =============================================================
paths:
...
  /api/v1/admin/audit/events:
    get:
      tags: [Audit - Admin Only]
      summary: List all audit events (paginated)
      description: Requires ADMIN role. Returns all banking events recorded by the audit service.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of audit events
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEventResponse'
                  totalElements:
                    type: integer
                    example: 4821
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit/events/{id}:
    get:
      tags: [Audit - Admin Only]
      summary: Get audit event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 42
      responses:
        '200':
          description: Audit event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit/ref/{referenceCode}:
    get:
      tags: [Audit - Admin Only]
      summary: Get audit events by business reference code
      description: Returns all audit entries related to a specific transaction, payment, or group reference code.
      parameters:
        - name: referenceCode
          in: path
          required: true
          schema:
            type: string
            example: "TXN-20240601-ABC123"
      responses:
        '200':
          description: Audit events for the reference code
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit/domain/{domain}:
    get:
      tags: [Audit - Admin Only]
      summary: Get audit events by domain
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EventDomain'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated audit events for the domain
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit/actor/{email}:
    get:
      tags: [Audit - Admin Only]
      summary: Get audit events by actor email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
            example: "jeanpierre@example.com"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: All audit events generated by the specified user
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/audit/stats:
    get:
      tags: [Audit - Admin Only]
      summary: Get audit statistics summary
      description: Returns aggregate counts by domain and event type — useful for admin dashboards.
      responses:
        '200':
          description: Audit statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalEvents:
                    type: integer
                    example: 4821
                  byDomain:
                    type: object
                    additionalProperties:
                      type: integer
                    example:
                      WALLET: 1200
                      TRANSACTION: 2100
                      MERCHANT: 800
                      SAVINGS: 721
                  byEventType:
                    type: object
                    additionalProperties:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

# =============================================================
# Tag descriptions
# =============================================================
...